requirements.txt
backcall==0.2.0
Brotli==1.0.9
certifi==2021.5.30
chardet==4.0.0
click==8.0.1
cycler==0.10.0
dash==1.20.0
dash-core-components==1.16.0
dash-html-components==1.1.3
dash-renderer==1.9.1
dash-table==4.11.3
DateTime==4.3
decorator==5.0.9
Flask==2.0.1
Flask-Compress==1.10.1
future==0.18.2
idna==2.10
ipython==7.25.0
ipython-genutils==0.2.0
itsdangerous==2.0.1
jedi==0.18.0
Jinja2==3.0.1
joblib==1.0.1
kiwisolver==1.3.1
MarkupSafe==2.0.1
matplotlib==3.4.2
matplotlib-inline==0.1.2
mysql-connector==2.2.9
nltk==3.6.2
numpy==1.21.0
oauthlib==3.1.1
pandas==1.2.5
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.3.0
plotly==5.1.0
prompt-toolkit==3.0.19
psycopg2==2.9.1
ptyprocess==0.7.0
Pygments==2.9.0
pyparsing==2.4.7
PySocks==1.7.1
python-dateutil==2.8.1
pytz==2021.1
regex==2021.4.4
requests==2.25.1
requests-oauthlib==1.3.0
scipy==1.7.0
seaborn==0.11.1
six==1.16.0
tenacity==7.0.0
textblob==0.15.3
tqdm==4.61.1
traitlets==5.0.5
tweepy==3.10.0
urllib3==1.26.5
wcwidth==0.2.5
Werkzeug==2.0.1
zope.interface==5.4.0